Upgrade to wgpu version 10
This is a work-in-progress to upgrade wgpu to version 10. The cargo
audit was failing because some of the dependencies of version 9 are no
longer maintained.

Several enums have been renamed in wgpu from version 9 to 10. There are
also some dependency resolution issues that were solved by switching to
the newest dependency resolver in cargo.

There are finally some other changes that I have not been able to figure
out. Maybe someone else can figure those out?